requirements.txt
h5py==3.0.0
keras-applications==1.0.8
keras-preprocessing==1.1.2

librosa==0.8.0

llvmlite==0.34.0
mir-eval==0.6

pandas==1.1.4

pyjwt==1.7.1
pyopenssl==19.1.0
pyparsing==2.4.7
pysocks==1.7.1
pysptk==0.1.18
pystoi==0.3.

pytorch-lightning==0.9.0
pytorch-ranger==0.1.1
pytz==2020.4
pyworld==0.2.12
pyyaml==5.3.1

resampy==0.2.2
resemblyzer==0.1.1.dev0
sacred==0.7.3
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0

six==1.15.0
smmap==3.0.4
soundfile==0.10.3.post1
tabulate==0.8.7
tensorboard==1.14.0

tensorboardx==2.1
tensorflow==1.15.0
tensorflow-estimator==1.14.0

torch==1.7.0+cu101
torch-stoi==0.0.1
torchaudio==0.7.0
torchvision==0.8.1+cu101
tqdm==4.51.0
vamp==1.1.0
webencodings==0.5.1==py37_1